change coverage to floa

+++ repository/repository.go
@@ -1,12 +1,12 @@
 package repository
 
 type ServiceCoverage struct {
-	ServiceName string `json:"service_name"`
-	Coverage    int    `json:"coverage"`
+	ServiceName string  `json:"service_name"`
+	Coverage    float32 `json:"coverage"`
 }
 
 type Repository interface {
-	UpdateServiceCoverage(serviceName string, coverage int)
+	UpdateServiceCoverage(serviceName string, coverage float32)
 	ListServiceCoverage() []ServiceCoverage
 }
 
@@ -14,7 +14,7 @@ type LocalRepository struct {
 	Services []ServiceCoverage
 }
 
-func (lr *LocalRepository) UpdateServiceCoverage(sn string, cov int) {
+func (lr *LocalRepository) UpdateServiceCoverage(sn string, cov float32) {
 	for i, v := range lr.Services {
 		if v.ServiceName == sn {
 			lr.Services[i].Coverage = cov
